+++ WebAPI/Controllers/TransactionsController.cs
@@ -1,15 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Serilog;
-using WebAPI.Models;
+using DbEntities.Models;
 using WebAPI.Services;
 
 namespace WebAPI.Controllers
