--- WebAPI/Models/Tag.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace WebAPI.Models
-{
-    public partial class Tag
-    {
-        public long Id { get; set; }
-        public string Title { get; set; }
-    }
-}
