--- WebAPI/Models/Transaction.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace WebAPI.Models
-{
-    public partial class Transaction
-    {
-        public long Id { get; set; }
-        public DateTime? TrTime { get; set; }
-        public decimal Amount { get; set; }
-        public string Details { get; set; }
-        public string CounterParty { get; set; }
-        public long? CategoryId { get; set; }
-
-        public virtual Category Category { get; set; }
-    }
-}
