+++ WebAPI/Services/BankStatmentReader.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using WebAPI.Models;
+using DbEntities.Models;
 
 namespace WebAPI.Services
 {
