+++ WebAPI/WebAPI.csproj
@@ -30,5 +30,9 @@
     <PackageReference Include="Serilog.Sinks.File" Version="4.1.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\DbEntities\DbEntities.csproj" />
+  </ItemGroup>
+
 
 </Project>
